requirements.txt
faiss-cpu
redis
openai
google-adk
PyPDF2